+++ changelog.md
@@ -80,6 +80,7 @@
 - ``parseOct`` and ``parseBin`` in parseutils now also support the ``maxLen`` argument similar to ``parseHexInt``
 - Added the proc ``flush`` for memory mapped files.
 - Added the ``MemMapFileStream``.
+- Added the proc ``sequtils.zipApply`` for mutating in a zipped iteration.
 
 ### Library changes